CORE-257: In Core Ethereum, don’t restore BRD/LCL nodes to AVAILABLE
We had restored BRD/LCL nodes to AVAILABLE right at the point of failure.  But then we repeatedly attempted to connect to the just-failed-but-now-AVAILABLE node.  Never moved on to find a working node.  This occurred when trying to connect to “Sam’s Parity Node” when outside of that node’s network.