CORE-265: In Core Ethereum, allow nodeConnect() to block LES
When we call nodeConnect() and the eventual Unix connect() reports EINPROGRESS, we’ll block for N (= 3) seconds.  This blocks the entire LES thread - which isn’t a good idea.  However, resolving this is onerous and the blocking shouldn’t occur often enough to be a problem.  Perhaps reconsider when CORE-205 is addressed.