CORE-297: Fix Ripple deserialize bug
The deserialize code uses the BRArray code (from support).
If the array size is increased due to a realloc the memory location
might change - so any function that is calling array_add needs to
be able to update the address of the pointer.  In my case I was passing
the array pointer to a function when I needed to pass the address of the
array pointer.

Tested by creating an array with capacity 1 and let it increase several times
in the fuction.  Unit tests all passed.