CORE-371: Announce Tokens at a high priority
When `ewmCreate()` is called, existing transactions and logs are reconstructed.  But, if EWM does not have a needed token, then reconstructed logs are simply dropped.  This change moves token announcements to the front of the EWM queue; it is a work-around.  The IOS/Android call sequence is ewmCreate which produces ewmSignalLog, ewmAnnounceToken and then ewmConnect.  By upping the priority the token announcements a bumped ahead of the signaled logs and then on connect the queue is drained.