CORE-398: Don't connect to 'community' nodes; be more flexible on transaction status
Regarding transaction status - if one node reports ERROR and a subsequent node reports PENDING or even QUEUED, move out of the ERROR state.  This happens commonly if a node is syncing; when syncing it rejects transaction submissions outright - we don't want to be stuck in ERROR in that case.  However, if multiple nodes report an ERROR (either because they are syncing or something like gas_too_low occurs) then we will go to ERROR.  We'll get out of ERROR when the transaction is actually included.  [A screw case exists where *all* connected nodes are syncing.]