CORE-407: Move CWM Callbacks from WalletManager to System
This is designed to resolve memory issues whereby in a callback the C memory has disappears, owing to a race condition as the callback can occur ‘late’.  System includes a statically allocated dictionary as [Int:Weak<System>]; if the callback does not find System, then the callback is skipped.

Also, removed listener properties in WalletManager, Wallet and Transfer; all events are generated at the System level.