CORE-466: Fix Ripple unit test
Fixed the unit test - we do not need to allow
for creating of account from private key. I was
able to create a new unit test to validate creating
of the transaction id

Remove unused code from the account file. It was only used
for testing and I now test a different way.