CORE-523: Address TSAN Issues (#1a)
Using the Crypto Demo in both P2P and API modes (and swapping between the two); only the above issues resolve TSAN.  And, probably only the `cryptoWalletManagerGetAddressScheme()` was enough and the other changes are not required.  However, given that the Demo does not use all interfaces, we'll need more analysis/testing/design-contraints.