CORE-447: In Core Ethereum, Update SYNC events
The transition from SYNCING -> CONNECTED signifies the completion of a 'sync'.  This transition was being signal well after the BDB query completed; it was signaled at the next periodic dispatch which is as much as 15 seconds later.  Signal it when the log and transaction queries have actually finished.
Also, mark the sync as completed no matter if the BDB query actually completed successfully; if it failed we'll just query again next time.