CORE-684: Always clone addresses getting/setting
When creating a wallet, account, transaction, or transfer
always copy the passed in BRRippleAddress so we are explicit
about ownership.  The caller still owne their passed in object
and must free.

When returning a BRRippleAddress from any of our objects, always
return a copy so the caller knows they must free it when done.

This will make dealing with addresses a bit of a pain but at least
it is clear what is happening.